Added "already banned" fail2ban check and further optimized the log output
A new debug control variable has been introduced to define whether the script shall output everything to STDOUT or only newly banned IP addresses. This helps to debug the function of the script when running as CRON job.

Additionally, we're using the fail2ban-client status command to check if the IP is already banned in order to optimize the log and debug outputs.