Fix fail2ban-client return value bug
fail2ban-client did not properly return 0 when it successfully banned an IP causing the script to not report the banning accurately. The script was extended to just ban the IP and then issue a check whether the IP was successfully banned. This, unfortunately, requires some more resources but since it only happens upon banning the performance impact is neglectable.