Fixing fail2ban-client return value bug - Debug output optimization
The last solution was not optimal and caused another bug during runtime. Reverted and changed the way of handling the fail2ban-client return value. Due to a bug in the fail2ban-client the return value comparison must be string-based. Strange.

Enhanced the debug output to report an error when there's actually an error occuring when banning the IP address.