Optimize Whitelisting
The whitelist check has been done too late in the code which caused the script to perform certain checks on IP addresses which did not need these checks to be performed on in the first place. The updated logic no longer performs these unnecessary checks, reduces potential errors in script execution as well as increases performance.